# packages/contracts/generated/control-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Symphony Encore Control API",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "in": "cookie",
        "name": "symphony_session",
        "type": "apiKey"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "Create an authenticated operator session",
        "tags": [
          "authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "auth_subject",
                  "password"
                ],
                "properties": {
                  "auth_subject": {
                    "maxLength": 512,
                    "minLength": 1,
                    "type": "string"
                  },
                  "password": {
                    "maxLength": 4096,
                    "minLength": 1,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "csrf_token",
                    "expires_at",
                    "operator"
                  ],
                  "properties": {
                    "csrf_token": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "expires_at": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "operator": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "auth_subject",
                        "capabilities",
                        "operator_id"
                      ],
                      "properties": {
                        "auth_subject": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "capabilities": {
                          "uniqueItems": true,
                          "type": "array",
                          "items": {
                            "minLength": 1,
                            "type": "string"
                          }
                        },
                        "operator_id": {
                          "minLength": 1,
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "current_version",
                        "details",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "current_version": {
                          "anyOf": [
                            {
                              "minLength": 1,
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "message": {
                          "minLength": 1,
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "current_version",
                        "details",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "current_version": {
                          "anyOf": [
                            {
                              "minLength": 1,
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "message": {
                          "minLength": 1,
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events/stream": {
      "get": {
        "operationId": "streamEvents",
        "summary": "Live durable event stream",
        "tags": [
          "control"
        ],
        "parameters": [
          {
            "schema": {
              "minimum": 0,
              "type": "integer"
            },
            "in": "query",
            "name": "after_cursor",
            "required": false
          },
          {
            "schema": {
              "pattern": "^(?:0|[1-9][0-9]*)$",
              "type": "string"
            },
            "in": "header",
            "name": "last-event-id",
            "required": false
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "text/event-stream": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "current_version",
                        "details",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "current_version": {
                          "anyOf": [
                            {
                              "minLength": 1,
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "message": {
                          "minLength": 1,
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "text/event-stream": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "current_version",
                        "details",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "current_version": {
                          "anyOf": [
                            {
                              "minLength": 1,
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "message": {
                          "minLength": 1,
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "text/event-stream": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "current_version",
                        "details",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "current_version": {
                          "anyOf": [
                            {
                              "minLength": 1,
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "message": {
                          "minLength": 1,
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/state": {
      "get": {
        "operationId": "getControlState",
        "summary": "Current orchestrator state",
        "tags": [
          "control"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "dispatch_enabled",
                    "mutations_enabled",
                    "service_run",
                    "version"
                  ],
                  "properties": {
                    "dispatch_enabled": {
                      "type": "boolean"
                    },
                    "mutations_enabled": {
                      "type": "boolean"
                    },
                    "service_run": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "id",
                        "service_version",
                        "started_at",
                        "status"
                      ],
                      "properties": {
                        "id": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "service_version": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "started_at": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "status": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "starting"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "recovering"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "ready"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "failed"
                              ]
                            }
                          ]
                        }
                      }
                    },
                    "version": {
                      "minLength": 1,
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "current_version",
                        "details",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "current_version": {
                          "anyOf": [
                            {
                              "minLength": 1,
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "message": {
                          "minLength": 1,
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "current_version",
                        "details",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "current_version": {
                          "anyOf": [
                            {
                              "minLength": 1,
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        },
                        "message": {
                          "minLength": 1,
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}